Make all file writes generic
This adds two new traits, `Truncate` and `Length`, which are two operations we needed
that required the use of `File`.

With these two traits, combined with `std::io::Read`, `std::io::Write` and `std::io::Seek`, other types can now be treated as files.

It is now possible to write to a `Cursor<Vec<u8>>`.

Signed-off-by: Serial <69764315+Serial-ATA@users.noreply.github.com>