MP4: Demote atom size mismatch error for `free`
If a `free` atom claims to be larger than the remainder of the stream, parsing will simply stop. This will now only be a `SizeMismatch` error in `Strict` mode. Invalid padding is a common issue in all tag formats due to buggy software, so it's better to work around it by default rather than discard the entire stream as invalid.